fixup! change behaviour to replicate Postgres

+++ src/sql/src/func.rs
@@ -1576,7 +1576,7 @@ lazy_static! {
                 params!(String, String, String) => VariadicFunc::Replace, 2087;
             },
             "right" => Scalar {
-                params!(String, Int64) => BinaryFunc::Right, 3061;
+                params!(String, Int32) => BinaryFunc::Right, 3061;
             },
             "round" => Scalar {
                 params!(Float32) => UnaryFunc::RoundFloat32, oid::FUNC_ROUND_F32_OID;
